moved the count down timer to its own component
