+++ public/style.css
@@ -6,7 +6,6 @@
          url('font/POCKC___.TTF') format('truetype');
 
     font-style: normal;
-
 }
 
 body,
