+++ views/Layout.jsx
@@ -1,6 +1,5 @@
 //const React = require('react');
 import React from 'react';
-import { CountdownCircleTimer } from 'react-countdown-circle-timer'
 
 import {
     Router,
@@ -29,6 +28,7 @@ import HTML from './components/HTML';
             import Main from './components/Main';
                 import Canvas from './components/Canvas';
                 import Canvas2 from './components/Canvas2';
+                import Countdown from './components/Countdown';
             import Footer from './components/Footer';
 
 
@@ -37,18 +37,6 @@ import { css } from "@emotion/core";
 import Loader from "react-spinners/PacmanLoader";
 
 const override = css;
-
-//for countdown timer
-const minuteSeconds = 60;
-const hourSeconds = 3600;
-const daySeconds = 86400;
-const stratTime = Date.now() / 1000;
-const endTime = 1653195600;
-const remainingTime = endTime - stratTime;
-const days = Math.ceil(remainingTime / daySeconds);
-const daysDuration = days * daySeconds;
-
-//end timer
 
 class Layout extends React.Component{
 //module.exports = React.createClass({
@@ -229,31 +217,6 @@ class Layout extends React.Component{
             </Body></HTML>);
         }
 
-        /*import About from './About';*/
-
-        const renderTime = (dimension, time) => {
-            return (
-                <div className="time-wrapper">
-                    <div className="time">{time}</div>
-                    <div>{dimension}</div>
-                </div>
-            );
-        };
-
-        const getTimeSeconds = (time) => (minuteSeconds - time) | 0;
-        const getTimeMinutes = (time) => ((time % hourSeconds) / minuteSeconds) | 0;
-        const getTimeHours = (time) => ((time % daySeconds) / hourSeconds) | 0;
-        const getTimeDays = (time) => (time / daySeconds) | 0;
-
-        const vars = {
-            isPlaying: true,
-            strokeWidth: 12,
-            colors: '#F00',
-            trailColor: '#000'
-        }
-
-        //const cWidth = mobileView ? (innerWidth) : (innerWidth/4);
-        //console.log('cWidth',cWidth,'mob',mobileView);
         return(
 			<HTML>
 
@@ -287,66 +250,7 @@ class Layout extends React.Component{
 
                 <Main height={innerHeight}>
 
-                    <div className='countdown'>
-                        <CountdownCircleTimer
-                            {...vars}
-                            duration={daysDuration}
-                            onComplete={() => ({ shouldRepeat: true, delay: 1 })}
-                            initialRemainingTime={remainingTime}
-                            size={innerWidth/4}
-                        >
-                            {({ elapsedTime, color }) => (
-                                <span style={{ color }}>
-                                    {renderTime("days", getTimeDays(daysDuration - elapsedTime))}
-                                </span>
-                            )}
-                        </CountdownCircleTimer>
-                        <CountdownCircleTimer
-                            {...vars}
-                            duration={daySeconds}
-                            initialRemainingTime={remainingTime % daySeconds}
-                            onComplete={(totalElapsedTime) => ({
-                                shouldRepeat: remainingTime - totalElapsedTime > hourSeconds
-                            })}
-                            size={innerWidth/4}
-                        >
-                            {({ elapsedTime, color }) => (
-                                <span style={{ color }}>
-                                    {renderTime("hours", getTimeHours(daySeconds - elapsedTime))}
-                                </span>
-                            )}
-                        </CountdownCircleTimer>
-                        <CountdownCircleTimer
-                            {...vars}
-                            duration={hourSeconds}
-                            initialRemainingTime={remainingTime % hourSeconds}
-                            onComplete={(totalElapsedTime) => ({
-                                shouldRepeat: remainingTime - totalElapsedTime > minuteSeconds
-                            })}
-                            size={innerWidth/4}
-                        >
-                            {({ elapsedTime, color }) => (
-                                <span style={{ color }}>
-                                    {renderTime("mins", getTimeMinutes(hourSeconds - elapsedTime))}
-                                </span>
-                            )}
-                        </CountdownCircleTimer>
-                        <CountdownCircleTimer
-                            {...vars}
-                            duration={minuteSeconds}
-                            initialRemainingTime={remainingTime % minuteSeconds}
-                            onComplete={(totalElapsedTime) => ({
-                                shouldRepeat: remainingTime - totalElapsedTime > 0
-                            })}
-                            size={innerWidth/4}
-                        >
-                            {({ elapsedTime, color }) => (
-                                <span style={{ color }}>
-                                    {renderTime("secs", getTimeSeconds(elapsedTime))}
-                                </span>
-                            )}
-                        </CountdownCircleTimer>
-                    </div>
+                    <Countdown />
                     <Canvas height={innerHeight} width={innerWidth}/>
                     <Canvas2 height={innerHeight} width={innerWidth} mobileView={mobileView} />
                 </Main>
